Bug fix
As for some Uniprot id, there is no suitable gene name, so when running line 163, it gets a 200 OK, but when running into  line 165, it gets no key 'geneName', so raises KeyError
The very first uniprot id identified is A0A1D6EIG7, I tried getting json by my browser and read the website of this id, the problem consists.
As far as I know, many other genes correspond this situation.
Similarly,  we also need to check the existence of the key 'gene' in the same line 165.
Also, in line 170, the name 'uniprot_id' is not defined, it may be 'uniprot_accession_number' instead?